вынес шаблон страницы в отдельный компонент
